Move sync field, sync list, net function list to identity it can reduce from 2 bytes (index of behaviour, index of sync field) to 1 byte (byte - uint) with only index of sync field that included in sync messages

+++ Scripts/GameApi/LiteNetLibSyncField.cs
@@ -116,7 +116,7 @@ namespace LiteNetLibManager
             return field.Value;
         }
 
-        internal override void Setup(LiteNetLibBehaviour behaviour, byte elementId)
+        internal override void Setup(LiteNetLibBehaviour behaviour, int elementId)
         {
             base.Setup(behaviour, elementId);
             if (onChange != null)
